Apache helper should bind to all IP addresses
Scenario: you are trying to replace a Mailman server, or just the web portion. In Puppet, you configure the new server to use the same hostname as the last one. But that hostname gets used to set up the IP binding in Apache, so Apache doesn't listen on any interfaces because your server doesn't own the production IP address yet. To enable pre-cutover debugging, make Apache listen to the wildcard interface.